requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
requests==2.32.3
urllib3==2.3.0
python-dotenv==1.0.0
azure-search-documents==11.4.0
num2words==0.5.12
openai==1.13.3
pdfminer==20191125
transformers==4.31.0
scikit-learn==1.3.0
plotly==5.16.1
matplotlib==3.7.2
pandas==2.1.2
azure-cognitiveservices-speech
tiktoken==0.7.0
httpx==0.27.2
